Move public functions to `ethutil.go`
